requirements.txt
google-genai==1.26.0
numpy==2.3.1
matplotlib==3.10.3